PR Comments #7.0 (Still needs testing)

+++ cbxp/cbxp.cpp
@@ -6,11 +6,10 @@
 #include "logger.hpp"
 
 cbxp_result_t* cbxp_extract(const char* control_block_name,
-                            const unsigned int control_block_name_length,
-                            const char* includes,
-                            const unsigned int includes_length,
-                            const char* filters,
-                            const unsigned int filters_length, bool debug) {
+                            const size_t control_block_name_length,
+                            const char* includes, const size_t includes_length,
+                            const char* filters, const size_t filters_length,
+                            bool debug) {
   std::string control_block_name_string;
   std::string includes_string_string;
   std::string filters_string_string;
@@ -42,9 +41,9 @@ cbxp_result_t* cbxp_extract(const char* control_block_name,
 }
 
 cbxp_result_t* cbxp_format(const char* control_block_name,
-                           const unsigned int control_block_name_length,
-                           const void* bytes_buffer,
-                           const unsigned int buffer_length, bool debug) {
+                           const size_t control_block_name_length,
+                           const void* bytes_buffer, const size_t buffer_length,
+                           bool debug) {
   std::string control_block_name_string;
 
   if (control_block_name != nullptr) {
