+++ cbxp/control_blocks/ascb.cpp
@@ -15,7 +15,7 @@
 
 namespace CBXP {
 nlohmann::json ASCB::get(const void* p_control_block,
-                         const unsigned int buffer_length) {
+                         const size_t buffer_length) {
   ASCB::checkBufferLength(buffer_length);
   nlohmann::json ascb_json = {};
   const ascb* p_ascb;
