+++ cbxp/control_blocks/asvt.hpp
@@ -32,8 +32,8 @@ namespace CBXP {
 
 class ASVT : public ControlBlock {
  public:
-  nlohmann::json get(const void* p_control_block      = nullptr,
-                     const unsigned int buffer_length = 0) override;
+  nlohmann::json get(const void* p_control_block = nullptr,
+                     const size_t buffer_length  = 0) override;
   explicit ASVT(const cbxp_options_t& cbxp_options)
       : ControlBlock("asvt", {"ascb"}, cbxp_options) {
     control_block_length_ = sizeof(asvt_t);
