+++ cbxp/control_blocks/control_block.hpp
@@ -42,9 +42,9 @@ class ControlBlock {
   bool skip_buffer_length_check_     = false;
 
  public:
-  void checkBufferLength(const unsigned int buffer_length) const;
-  virtual nlohmann::json get(const void* p_control_block      = nullptr,
-                             const unsigned int buffer_length = 0) = 0;
+  void checkBufferLength(const size_t buffer_length) const;
+  virtual nlohmann::json get(const void* p_control_block = nullptr,
+                             const size_t buffer_length  = 0) = 0;
   explicit ControlBlock(const std::string& name,
                         const std::vector<std::string>& includables,
                         const cbxp_options_t& cbxp_options)
