+++ cbxp/control_blocks/ecvt.cpp
@@ -12,7 +12,7 @@
 
 namespace CBXP {
 nlohmann::json ECVT::get(const void* p_control_block,
-                         const unsigned int buffer_length) {
+                         const size_t buffer_length) {
   ECVT::checkBufferLength(buffer_length);
   const struct ecvt* p_ecvt;
   nlohmann::json ecvt_json = {};
