+++ cbxp/control_blocks/ecvt.hpp
@@ -9,8 +9,8 @@ namespace CBXP {
 
 class ECVT : public ControlBlock {
  public:
-  nlohmann::json get(const void* p_control_block      = nullptr,
-                     const unsigned int buffer_length = 0) override;
+  nlohmann::json get(const void* p_control_block = nullptr,
+                     const size_t buffer_length  = 0) override;
   explicit ECVT(const cbxp_options_t& cbxp_options)
       : ControlBlock("ecvt", {}, cbxp_options) {
     control_block_length_ = sizeof(struct ecvt);
