+++ cbxp/control_blocks/oucb.cpp
@@ -14,7 +14,7 @@
 
 namespace CBXP {
 nlohmann::json OUCB::get(const void* p_control_block,
-                         const unsigned int buffer_length) {
+                         const size_t buffer_length) {
   OUCB::checkBufferLength(buffer_length);
   const oucb_t* p_oucb;
   nlohmann::json oucb_json = {};
